requirements.txt
Flask
flask-cors
boto3
python-dotenv
pytest
pytest-mock
plaid-python
pytest-cov 
coverage
